+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/eventregistration/EventRegistrationModule.kt
@@ -2,8 +2,6 @@ package de.rki.coronawarnapp.eventregistration
 
 import dagger.Binds
 import dagger.Module
-import de.rki.coronawarnapp.eventregistration.checkins.CheckInsTransformer
-import de.rki.coronawarnapp.eventregistration.checkins.DefaultCheckInsTransformer
 import de.rki.coronawarnapp.eventregistration.checkins.download.DownloadedCheckInsRepo
 import de.rki.coronawarnapp.eventregistration.checkins.download.FakeDownloadedCheckInsRepo
 import de.rki.coronawarnapp.eventregistration.storage.repo.DefaultTraceLocationRepository
@@ -11,10 +9,6 @@ import de.rki.coronawarnapp.eventregistration.storage.repo.TraceLocationReposito
 
 @Module
 abstract class EventRegistrationModule {
-
-    @Binds
-    abstract fun checkInsTransformer(transformer: DefaultCheckInsTransformer): CheckInsTransformer
-
     @Binds
     abstract fun traceLocationRepository(defaultTraceLocationRepo: DefaultTraceLocationRepository):
         TraceLocationRepository
