+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/eventregistration/checkins/CheckInsTransformerTest.kt
@@ -7,9 +7,9 @@ import org.joda.time.Instant
 import org.junit.jupiter.api.Test
 import testhelpers.BaseTest
 
-class DefaultCheckInsTransformerTest : BaseTest() {
+class CheckInsTransformerTest : BaseTest() {
 
-    private val checkInTransformer = DefaultCheckInsTransformer()
+    private val checkInTransformer = CheckInsTransformer()
 
     @Test
     fun `transform check-ins`() {
